web/apps/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv 